Move YouTubeMusic fragment to a saparate package
