+++ app/src/main/java/com/yurii/youtubemusic/screens/youtube/YouTubeMusicsFragment.kt
@@ -1,10 +1,9 @@
-package com.yurii.youtubemusic
+package com.yurii.youtubemusic.screens.youtube
 
 import android.annotation.SuppressLint
 import android.os.Bundle
 import android.view.View
 import androidx.coordinatorlayout.widget.CoordinatorLayout
-import androidx.databinding.ViewDataBinding
 import androidx.fragment.app.activityViewModels
 import androidx.fragment.app.viewModels
 import com.google.api.services.youtube.model.Playlist
@@ -16,6 +15,7 @@ import androidx.recyclerview.widget.LinearLayoutManager
 import com.google.android.gms.auth.api.signin.GoogleSignInAccount
 import com.google.android.material.bottomnavigation.BottomNavigationView
 import com.google.api.services.youtube.model.PlaylistListResponse
+import com.yurii.youtubemusic.R
 import com.yurii.youtubemusic.playlists.PlayListsDialogInterface
 import com.yurii.youtubemusic.models.VideoItem
 import com.yurii.youtubemusic.services.downloader.Progress
@@ -28,9 +28,6 @@ import com.yurii.youtubemusic.ui.SelectCategoriesDialog
 import com.yurii.youtubemusic.adapters.VideosListAdapter
 import com.yurii.youtubemusic.services.downloader.ServiceConnection
 import com.yurii.youtubemusic.screens.main.MainActivityViewModel
-import com.yurii.youtubemusic.viewmodels.VideosLoader
-import com.yurii.youtubemusic.viewmodels.YouTubeMusicViewModel
-import com.yurii.youtubemusic.viewmodels.YouTubeViewModelFactory
 import java.lang.Exception
 import java.lang.IllegalArgumentException
 
@@ -42,7 +39,7 @@ class YouTubeMusicsFragment : TabFragment<FragmentYouTubeMusicsBinding>(
 ), VideosLoader {
     private val mainActivityViewModel: MainActivityViewModel by activityViewModels()
     private val viewModel: YouTubeMusicViewModel by viewModels {
-        YouTubeViewModelFactory(requireActivity().application, getGoogleSignInAccount(), ServiceLocator.providePreferences(requireContext()))
+        Injector.provideYouTubeMusicViewModel(requireContext(), getGoogleSignInAccount())
     }
 
     private lateinit var videosListAdapter: VideosListAdapter
